Fixed the bugs that caused problems in rsxglgears.
The incorrect rendering in the gears example seems to have been caused mostly
by RSXGL setting the default state of glDepthFunc() to GL_ALWAYS instead of GL_LESS
as it should be.

Parts of state.cc pertaining to front and back face culling got a wash and a rinse
as well.