FBO demo does something more useful, now.
It renders an image of a spinning cube to a texture, which is then displayed on another spinning cube in the main framebuffer. The colors of the texture on the inner, FBO-targed cube are obviously wrong, but I'm not sure where the incorrect color channel swizzling is happening. "w00t" nonetheless, given the amount of time it's taken me to implement OpenGL's FBO API.