Instanced rendering.
Only glDrawElementsInstanced has been tested. Vertex program assembler
needs to handle texture opcodes in order for this to be useful.

GLSL defines a gl_InstanceID variable for vertex programs to access
the current instance being rendered. NVidia cgc compiler doesn't
implement this for the vp40 profile supported by the RSX, but
reserves the variable name nonetheless. Therefore RSXGL looks instead
for a vertex program constant called rsxgl_InstanceID and assigns the
current instance ID to it instead. It may also be necessary for the
instance id variable to be a vertex program attribute instead of a
constant in order to successfully or efficiently perform texture reads
to compute outputs per-vertex.